+++ eras/shelley/impl/src/Cardano/Ledger/Shelley/TxOut.hs
@@ -34,8 +34,11 @@ import Cardano.Ledger.Binary (
   FromCBOR (..),
   Interns (..),
   ToCBOR (..),
+  TokenType (..),
+  decodeMemPack,
   decodeRecordNamed,
   encodeListLen,
+  peekTokenType,
  )
 import Cardano.Ledger.Compactible (Compactible (CompactForm, fromCompact, toCompact))
 import Cardano.Ledger.Core
@@ -45,8 +48,9 @@ import Cardano.Ledger.Shelley.PParams ()
 import Cardano.Ledger.Val (Val)
 import Control.DeepSeq (NFData (rnf))
 import Data.Aeson (KeyValue, ToJSON (..), object, pairs, (.=))
-import Data.ByteString.Short (ShortByteString, pack)
+import qualified Data.ByteString.Short as SBS (ShortByteString, pack)
 import Data.Maybe (fromMaybe)
+import Data.MemPack
 import Data.Proxy (Proxy (..))
 import Data.Word (Word8)
 import GHC.Stack (HasCallStack)
@@ -57,6 +61,23 @@ data ShelleyTxOut era = TxOutCompact
   { txOutCompactAddr :: {-# UNPACK #-} !CompactAddr
   , txOutCompactValue :: !(CompactForm (Value era))
   }
+
+-- | This instance uses a zero Tag for forward compatibility in binary representation with TxOut
+-- instances for future eras
+instance (Era era, MemPack (CompactForm (Value era))) => MemPack (ShelleyTxOut era) where
+  packedByteCount = \case
+    TxOutCompact cAddr cValue ->
+      packedTagByteCount + packedByteCount cAddr + packedByteCount cValue
+  {-# INLINE packedByteCount #-}
+  packM = \case
+    TxOutCompact cAddr cValue ->
+      packTagM 0 >> packM cAddr >> packM cValue
+  {-# INLINE packM #-}
+  unpackM =
+    unpackTagM >>= \case
+      0 -> TxOutCompact <$> unpackM <*> unpackM
+      n -> unknownTagM @(ShelleyTxOut era) n
+  {-# INLINE unpackM #-}
 
 instance EraTxOut ShelleyEra where
   type TxOut ShelleyEra = ShelleyTxOut ShelleyEra
@@ -146,9 +167,19 @@ instance (Era era, DecCBOR (CompactForm (Value era))) => DecCBOR (ShelleyTxOut e
       cAddr <- decCBOR
       TxOutCompact cAddr <$> decCBOR
 
-instance (Era era, DecCBOR (CompactForm (Value era))) => DecShareCBOR (ShelleyTxOut era) where
+instance
+  ( Era era
+  , MemPack (CompactForm (Value era))
+  , DecCBOR (CompactForm (Value era))
+  ) =>
+  DecShareCBOR (ShelleyTxOut era)
+  where
   type Share (ShelleyTxOut era) = Interns (Credential 'Staking)
-  decShareCBOR _ = decCBOR
+  decShareCBOR _ = do
+    peekTokenType >>= \case
+      TypeBytes -> decodeMemPack
+      TypeBytesIndef -> decodeMemPack
+      _ -> decCBOR
 
 instance (Era era, EncCBOR (CompactForm (Value era))) => ToCBOR (ShelleyTxOut era) where
   toCBOR = toEraCBOR @era
@@ -168,9 +199,9 @@ toTxOutPair (ShelleyTxOut !addr !amount) =
 
 -- a ShortByteString of the same length as the ADDRHASH
 -- used to calculate heapWords
-packedADDRHASH :: ShortByteString
+packedADDRHASH :: SBS.ShortByteString
 packedADDRHASH =
-  pack $
+  SBS.pack $
     replicate
       (fromIntegral (1 + 2 * HS.sizeHash (Proxy :: Proxy ADDRHASH)))
       (1 :: Word8)