+++ libs/cardano-ledger-core/testlib/Test/Cardano/Ledger/Core/Binary/RoundTrip.hs
@@ -216,6 +216,11 @@ roundTripCoreEraTypesSpec = do
     roundTripAnnEraSpec @era @(TxWits era)
     roundTripAnnEraSpec @era @(TxBody era)
     roundTripAnnEraSpec @era @(Tx era)
+    prop ("MemPack/CBOR Roundtrip " <> show (typeRep $ Proxy @(TxOut era))) $
+      roundTripRangeExpectation @(TxOut era)
+        (mkTrip encodeMemPack decNoShareCBOR)
+        (eraProtVerLow @era)
+        (eraProtVerHigh @era)
   describe "Core State Types" $ do
     roundTripAnnEraSpec @era @BootstrapWitness
     roundTripShareEraSpec @era @SnapShots